backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-jose[cryptography]
passlib[bcrypt]
python-magic
python-magic-bin
Pillow>=10.1.0
imageio
imageio-ffmpeg
sqlalchemy
alembic
psycopg2-binary
pydantic[email]
pydantic-settings
python-dotenv
redis
email-validator
cryptography
httpx
aiofiles
itsdangerous
bcrypt==3.2.2
typing-extensions